@@ -8,38 +8,26 @@
   "homepage": "https://github.com/shellscape/loglevelnext",
   "main": "lib/index.js",
   "engines": {
-    "node": ">= 8.16.0"
+    "node": ">= 12.22.1"
   },
   "scripts": {
-    "ci:coverage": "nyc npm run test && nyc report --reporter=text-lcov > coverage.lcov",
-    "ci:lint": "npm run lint && npm run security",
-    "ci:lint:commits": "commitlint --from=${CIRCLE_BRANCH} --to=${CIRCLE_SHA1}",
-    "ci:test": "npm run test",
+    "build": "tsc --project tsconfig.json",
+    "ci:coverage": "nyc pnpm test && nyc report --reporter=text-lcov > coverage.lcov",
     "commitlint": "commitlint",
     "commitmsg": "commitlint -e $GIT_PARAMS",
-    "lint": "eslint --fix --cache lib test",
+    "lint": "pnpm lint:docs && pnpm lint:json && pnpm lint:package && pnpm lint:js",
     "lint-staged": "lint-staged",
-    "security": "npm audit",
-    "test": "ava"
+    "lint:docs": "prettier --write .github/**/*.md **/README.md",
+    "lint:js": "eslint --fix --cache src test",
+    "lint:json": "prettier --write tsconfig.*.json",
+    "lint:package": "prettier --write **/package.json --plugin=prettier-plugin-package",
+    "security": "pnpm audit --audit-level high",
+    "test": "FORCE_COLOR=3 ava"
   },
   "files": [
-    "lib/",
-    "types/",
+    "dist",
     "README.md"
   ],
-  "dependencies": {},
-  "devDependencies": {
-    "@commitlint/cli": "^8.2.0",
-    "@commitlint/config-conventional": "^8.2.0",
-    "ava": "^2.4.0",
-    "chalk": "^4.0.0",
-    "eslint-config-shellscape": "^2.1.0",
-    "lint-staged": "^10.1.3",
-    "nyc": "^15.0.1",
-    "pre-commit": "^1.2.2",
-    "sinon": "^7.2.3"
-  },
-  "types": "types/index.d.ts",
   "keywords": [
     "browser",
     "console",
@@ -58,10 +46,48 @@
     "trace",
     "warn"
   ],
-  "pre-commit": "lint-staged",
+  "devDependencies": {
+    "@commitlint/cli": "12.1.4",
+    "@commitlint/config-conventional": "12.1.4",
+    "@types/node": "^15.12.2",
+    "@types/sinon": "^10.0.2",
+    "ava": "^3.15.0",
+    "chalk": "^4.0.0",
+    "eslint-config-shellscape": "^4.0.1",
+    "husky": "^4.3.8",
+    "lint-staged": "11.0.0",
+    "nyc": "^15.0.1",
+    "pre-commit": "^1.2.2",
+    "sinon": "11.1.1",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
+  },
+  "types": "types/index.d.ts",
+  "ava": {
+    "extensions": [
+      "ts"
+    ],
+    "require": [
+      "ts-node/register"
+    ],
+    "files": [
+      "test/*.ts"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
   "lint-staged": {
-    "*.js": [
+    "*.{ts,js}": [
       "eslint --fix"
+    ],
+    "*.{json,md,yml,yaml}": [
+      "prettier --write"
+    ],
+    "*package.json": [
+      "prettier --write --plugin=prettier-plugin-package"
     ]
   }
 }
