@@ -1,5 +1,5 @@
 /*
-  Copyright © 2018 Andrew Powell
+  Copyright © 2021 Andrew Powell
 
   This Source Code Form is subject to the terms of the Mozilla Public
   License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -9,33 +9,61 @@
   included in all copies or substantial portions of this Source Code Form.
 */
 
+import { LogLevel } from './LogLevel';
+
+interface BindTarget {
+  [key: string]: Function;
+}
+
+export interface FactoryLevels extends Record<Uppercase<string>, number> {
+  SILENT: number;
+}
+
+export interface Factory {
+  [key: string]: any;
+  bindMethod: (obj: BindTarget, methodName: string) => any;
+  distillLevel: (level: number | string) => any;
+
+  levelValid: (level: number | string) => boolean;
+  levels: FactoryLevels;
+  logger: LogLevel;
+  make: (methodName: string) => Function;
+  methods?: string[];
+
+  replaceMethods: (logLevel: number | string) => void;
+}
+
 const noop = () => {};
 const levels = Symbol('log-levels');
 const instance = Symbol('log-instance');
 
-module.exports = class MethodFactory {
-  constructor(logger) {
-    this[instance] = logger;
-    this[levels] = {
-      TRACE: 0,
-      DEBUG: 1,
-      INFO: 2,
-      WARN: 3,
-      ERROR: 4,
-      SILENT: 5
-    };
+/* eslint-disable sort-keys */
+export const MethodFactoryLevels: FactoryLevels = {
+  TRACE: 0,
+  DEBUG: 1,
+  INFO: 2,
+  WARN: 3,
+  ERROR: 4,
+  SILENT: 5
+};
+/* eslint-enable sort-keys */
+
+export class MethodFactory implements Factory {
+  constructor(logger?: LogLevel) {
+    (this as any)[instance] = logger;
+    (this as any)[levels] = MethodFactoryLevels;
   }
 
-  get levels() {
-    return this[levels];
+  get levels(): typeof MethodFactoryLevels {
+    return (this as any)[levels];
   }
 
-  get logger() {
-    return this[instance];
+  get logger(): LogLevel {
+    return (this as any)[instance];
   }
 
-  set logger(logger) {
-    this[instance] = logger;
+  set logger(logger: LogLevel) {
+    (this as any)[instance] = logger;
   }
 
   get methods() {
@@ -45,7 +73,7 @@ module.exports = class MethodFactory {
   }
 
   // eslint-disable-next-line class-methods-use-this
-  bindMethod(obj, methodName) {
+  bindMethod(obj: BindTarget, methodName: string) {
     const method = obj[methodName];
     if (typeof method.bind === 'function') {
       return method.bind(obj);
@@ -62,7 +90,7 @@ module.exports = class MethodFactory {
     }
   }
 
-  distillLevel(level) {
+  distillLevel(level: number | string) {
     let result = level;
 
     if (typeof result === 'string' && typeof this.levels[result.toUpperCase()] !== 'undefined') {
@@ -73,10 +101,10 @@ module.exports = class MethodFactory {
       return result;
     }
 
-    return false;
+    return null;
   }
 
-  levelValid(level) {
+  levelValid(level: number | string) {
     if (typeof level === 'number' && level >= 0 && level <= this.levels.SILENT) {
       return true;
     }
@@ -91,22 +119,23 @@ module.exports = class MethodFactory {
    * console to become available.
    */
   // eslint-disable-next-line class-methods-use-this
-  make(methodName) {
+  make(methodName: string): Function {
     /* eslint-disable no-console */
-    if (typeof console[methodName] !== 'undefined') {
-      return this.bindMethod(console, methodName);
+    const target = console as unknown as BindTarget;
+    if (typeof target[methodName] !== 'undefined') {
+      return this.bindMethod(target, methodName);
     } else if (typeof console.log !== 'undefined') {
-      return this.bindMethod(console, 'log');
+      return this.bindMethod(target, 'log');
     }
 
     /* eslint-enable no-console */
     return noop;
   }
 
-  replaceMethods(logLevel) {
+  replaceMethods(logLevel: number | string) {
     const level = this.distillLevel(logLevel);
 
-    if (level == null) {
+    if (level === null) {
       throw new Error(`loglevelnext: replaceMethods() called with invalid level: ${logLevel}`);
     }
 
@@ -125,4 +154,4 @@ module.exports = class MethodFactory {
     // Define log.log as an alias for log.debug
     this.logger.log = this.logger.debug;
   }
-};
+}
