@@ -1,31 +1,37 @@
-const test = require('ava');
-const chalk = require('chalk');
-const sinon = require('sinon');
+import chalk from 'chalk';
+import sinon from 'sinon';
+import test from 'ava';
 
-const log = require('../lib');
+import log from '../src';
 
 const sandbox = sinon.createSandbox();
 const spyMethods = Object.keys(log.levels)
   .map((key) => key.toLowerCase())
   .filter((key) => key !== 'silent');
 
+interface ConsoleIndex {
+  [key: string]: Function;
+}
+
+const con = console as unknown as ConsoleIndex;
+
 test.before(() => {
   for (const method of spyMethods) {
-    if (console[method]) {
+    if (con[method]) {
       // curiously, node 6 doesn't have console.debug
-      sandbox.spy(console, method);
+      sandbox.spy(con, method as any);
     }
   }
-  sandbox.spy(console, 'log');
+  sandbox.spy(con, 'log');
 });
 
 test.afterEach(() => {
   for (const method of spyMethods) {
-    if (console[method]) {
-      console[method].resetHistory();
+    if (con[method]) {
+      (con[method] as sinon.SinonSpy).resetHistory();
     }
   }
-  console.log.resetHistory();
+  (con.log as sinon.SinonSpy).resetHistory();
 });
 
 test.after(() => {
@@ -62,11 +68,12 @@ for (const name of Object.keys(log.levels)) {
         expected = 0;
       }
 
-      if (method === 'error' && name === 'TRACE' && expected === 1) {
-        expected = 2;
-      }
+      // This was true for earlier node versions
+      // if (method === 'error' && name === 'TRACE' && expected === 1) {
+      //   expected = 2;
+      // }
 
-      t.is(console[method].callCount, expected);
+      t.is((con[method] as sinon.SinonSpy).callCount, expected);
     }
   });
 }
@@ -83,10 +90,11 @@ test.serial('enable() sets TRACE', (t) => {
 
 test.serial('throws on invalid levels', (t) => {
   t.throws(() => {
-    log.level = null;
+    (log as any).level = null;
   });
   t.throws(() => {
-    log.level = undefined; // eslint-disable-line no-undefined
+    // eslint-disable-next-line no-undefined
+    (log as any).level = undefined;
   });
   t.throws(() => {
     log.level = -1;
